public/data/scripts/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
Pillow==10.2.0
pandas==2.2.0
numpy==1.26.3
openai==1.12.0
aiohttp==3.9.3
tqdm==4.66.1
python-slugify==8.0.4
markdown==3.5.2
jinja2==3.1.3
textblob==0.17.1
colorthief==0.2.1 